Separated service locator definition from macros
