+++ src/entities/tools/FreeCam.h
@@ -2,8 +2,8 @@
 #define A_DARK_DISCOMFORT_FREECAM_H
 
 #include "../../engine/entity/Entity.h"
+#include "../../ServiceLocator.h"
 #include <Camera3D.hpp>
-#include "../../utils/ServiceLocator.h"
 
 class FreeCam : public Entity
 {
